move ThemeProvider into themeContext folder and link to GlobalStyle and themes
